Version 1.8
➢ Added compatibility with minecraft 1.21.5;
➢ Bug Fix of DayChallenge
➢ Bug Fix of DayChallenge;
➢ Bug Fix various;
➢ Code optimization;
