+++ VanillaChallengesSource/pom.xml
@@ -2,7 +2,7 @@
     <modelVersion>4.0.0</modelVersion>
     <groupId>io.eliotesta98.VanillaChallenges</groupId>
     <artifactId>VanillaChallengesSource</artifactId>
-    <version>1.7.3</version>
+    <version>1.8</version>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8
@@ -157,7 +157,7 @@
         <dependency>
             <groupId>org.spigotmc</groupId>
             <artifactId>spigot-api</artifactId>
-            <version>1.21.4-R0.1-SNAPSHOT</version>
+            <version>1.21.5-R0.1-SNAPSHOT</version>
             <scope>provided</scope>
         </dependency>
         <!--  h2  -->
@@ -218,7 +218,7 @@
         <dependency>
             <groupId>de.tr7zw</groupId>
             <artifactId>item-nbt-api</artifactId>
-            <version>2.14.1-SNAPSHOT</version>
+            <version>2.15.0</version>
             <scope>compile</scope>
             <exclusions>
                 <exclusion>
