+++ VanillaChallengesSource/src/main/java/io/eliotesta98/VanillaChallenges/Database/Database.java
@@ -1,91 +1,850 @@
 package io.eliotesta98.VanillaChallenges.Database;
 
+import io.eliotesta98.VanillaChallenges.Core.Main;
 import io.eliotesta98.VanillaChallenges.Database.Objects.Challenger;
 import io.eliotesta98.VanillaChallenges.Database.Objects.DailyWinner;
 import io.eliotesta98.VanillaChallenges.Database.Objects.PlayerStats;
 import io.eliotesta98.VanillaChallenges.Utils.Challenge;
-import java.util.ArrayList;
-import java.util.List;
+import io.eliotesta98.VanillaChallenges.Utils.ColorUtils;
+import io.eliotesta98.VanillaChallenges.Utils.MoneyUtils;
+import jdk.jfr.internal.LogLevel;
+import jdk.jfr.internal.LogTag;
+import jdk.jfr.internal.Logger;
+import org.bukkit.Bukkit;
+import org.bukkit.ChatColor;
+import org.bukkit.configuration.file.YamlConfiguration;
 
-public interface Database {
+import java.io.File;
+import java.io.IOException;
+import java.sql.*;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.util.*;
+import java.util.Date;
 
-    void initialize(String AbsolutePath);
+public abstract class Database {
 
-    String insertDailyChallenges();
+    private final List<Challenger> playerPoints = new ArrayList<>();
+    private final List<Challenge> challenges = new ArrayList<>();
+    private final List<DailyWinner> dailyWinners = new ArrayList<>();
+    private final List<Challenger> topYesterday = new ArrayList<>();
+    private final List<Challenger> oldPoints = new ArrayList<>();
+    private final List<PlayerStats> stats = new ArrayList<>();
+    private String prefix = "";
+    private Connection connection;
 
-    void insertChallengeEvent(String challengeName, int time);
+    public abstract void createConnection(String absolutePath);
 
-    void deleteChallengeWithName(String challengeName);
+    public void initialize() {
+        try {
+            PreparedStatement preparedStatement = connection.prepareStatement(
+                    "CREATE TABLE IF NOT EXISTS " + prefix + "Challenge (`NomeChallenge` VARCHAR(100) NOT NULL PRIMARY KEY, `TimeResume` INT(15) NOT NULL);");
+            preparedStatement.executeUpdate();
+            preparedStatement.close();
+            preparedStatement = connection.prepareStatement(
+                    "CREATE TABLE IF NOT EXISTS " + prefix + "Challenger (`PlayerName` VARCHAR(100) NOT NULL PRIMARY KEY, `Points` INT(15) NOT NULL);");
+            preparedStatement.executeUpdate();
+            preparedStatement.close();
+            preparedStatement = connection.prepareStatement(
+                    "CREATE TABLE IF NOT EXISTS " + prefix + "ChallengerEvent (`PlayerName` VARCHAR(100) NOT NULL PRIMARY KEY, `Points` INT(15) NOT NULL);");
+            preparedStatement.executeUpdate();
+            preparedStatement.close();
+            preparedStatement = connection.prepareStatement(
+                    "CREATE TABLE IF NOT EXISTS " + prefix + "DailyWinner (`ID` INT(100) NOT NULL AUTO_INCREMENT PRIMARY KEY, `NomeChallenge` VARCHAR(100) NOT NULL, `PlayerName` VARCHAR(100) NOT NULL, `Reward` VARCHAR(100) NOT NULL);");
+            preparedStatement.executeUpdate();
+            preparedStatement.close();
+            preparedStatement = connection.prepareStatement(
+                    "CREATE TABLE IF NOT EXISTS " + prefix + "TopYesterday (`ID` INT(100) NOT NULL AUTO_INCREMENT PRIMARY KEY, `PlayerName` VARCHAR(100) NOT NULL, `Points` INT(15) NOT NULL);");
+            preparedStatement.executeUpdate();
+            preparedStatement.close();
+            preparedStatement = connection.prepareStatement(
+                    "CREATE TABLE IF NOT EXISTS " + prefix + "Statistic (`PlayerName` VARCHAR(100) NOT NULL PRIMARY KEY, `NumberVictories` INT(10) NOT NULL, `NumberFirstPlace` INT(10) NOT NULL, `NumberSecondPlace` INT(10) NOT NULL, `NumberThirdPlace` INT(10) NOT NULL);");
+            preparedStatement.executeUpdate();
+            preparedStatement.close();
+        } catch (SQLException e) {
+            Main.instance.getServer().getConsoleSender().sendMessage("§cError Database not connected!");
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+            Main.instance.onDisable();
+        }
+        selectAllChallenges();
+        selectAllChallengers();
+        selectAllStats();
+        selectAllDailyWinners();
+        selectAllChallengersTopYesterday();
+        selectAllOldChallengers();
+    }
 
-    void loadPlayersPoints();
+    public void setConnection(Connection connection) {
+        this.connection = connection;
+    }
 
-    void disconnect();
+    public Connection getConnection() {
+        return this.connection;
+    }
 
-    ArrayList<PlayerStats> getAllPlayerStats();
+    public void setPrefix(String prefix) {
+        this.prefix = prefix;
+    }
 
-    void insertPlayerStat(PlayerStats playerStats);
+    public List<Challenge> getChallenges() {
+        return challenges;
+    }
 
-    void deletePlayerStatWithPlayerName(String playerName);
+    public List<PlayerStats> getAllStats() {
+        return stats;
+    }
 
-    void updatePlayerStat(PlayerStats playerStats);
+    public List<Challenger> getTopYesterday() {
+        return topYesterday;
+    }
 
-    ArrayList<Challenger> getAllChallengersTopYesterday();
+    public List<Challenger> getPlayerPoints() {
+        return playerPoints;
+    }
 
-    boolean isPresent(String playerName);
+    public List<Challenger> getOldPoints() {
+        return oldPoints;
+    }
 
-    void updateChallenger(String playerName, long points);
+    public List<DailyWinner> getDailyWinners() {
+        return dailyWinners;
+    }
 
-    void insertChallenger(String playerName, long point);
+    public void addPlayerPoints(Challenger challenger) {
+        this.playerPoints.add(challenger);
+    }
 
-    void updateDailyWinner(DailyWinner dailyWinner);
+    public void addChallenge(Challenge challenge) {
+        this.challenges.add(challenge);
+    }
 
-    void removeTopYesterday();
+    public void addChallenge(Challenge challenge, int index) {
+        this.challenges.add(index, challenge);
+    }
 
-    void saveTopYesterday(List<Challenger> newTopYesterday);
+    public void addDailyWinner(DailyWinner dailyWinner) {
+        this.dailyWinners.add(dailyWinner);
+    }
 
-    int lastDailyWinnerId();
+    public void addOldPoints(Challenger challenger) {
+        this.oldPoints.add(challenger);
+    }
 
-    void backupDb(int numberOfBackupedFile);
+    public void addTopYesterday(Challenger challenger) {
+        this.topYesterday.add(challenger);
+    }
 
-    void insertDailyWinner(DailyWinner dailyWinner);
+    public void addStat(PlayerStats playerStats) {
+        this.stats.add(playerStats);
+    }
 
-    void clearChallengers();
+    public void removeStat(String playerName) {
+        this.stats.removeIf((PlayerStats playerStat) -> playerStat.getPlayerName().equalsIgnoreCase(playerName));
+    }
 
-    void updateChallenge(String nomeChallenge,int timeResume);
+    public void removeChallenge(String challengeName) {
+        this.challenges.removeIf((Challenge challenge) -> challenge.getChallengeName().equalsIgnoreCase(challengeName));
+    }
 
-    ArrayList<DailyWinner> getAllDailyWinners();
+    public void removePlayer(String playerName) {
+        this.playerPoints.removeIf((Challenger challenger) -> challenger.getNomePlayer().equalsIgnoreCase(playerName));
+    }
 
-    void deleteDailyWinnerWithId(int id);
+    public void removeDailyWinner(int id) {
+        dailyWinners.removeIf((DailyWinner dailyWinner) -> dailyWinner.getId() == id);
+    }
 
-    PlayerStats getStatsPlayer(String playerName);
+    public void updatePlayer(String playerName, long points) {
+        removePlayer(playerName);
+        addPlayerPoints(new Challenger(playerName, points));
+    }
 
-    ArrayList<PlayerStats> getTopVictories();
+    public void updateChallengeTime(String challengeName, int time) {
+        for (Challenge challenge : challenges) {
+            if (challenge.getChallengeName().equalsIgnoreCase(challengeName)) {
+                challenge.setTimeChallenge(time);
+            }
+        }
+    }
 
-    ArrayList<PlayerStats> getTopFirstPlace();
+    public void selectAllChallenges() {
+        ResultSet resultSet;
+        try {
+            PreparedStatement preparedStatement = connection.prepareStatement("SELECT * FROM " + prefix + "Challenge");
+            resultSet = preparedStatement.executeQuery();
+            while (resultSet.next()) {
+                final Challenge challengeDB = new Challenge();
+                challengeDB.setTimeChallenge(resultSet.getInt("TimeResume"));
+                challengeDB.setChallengeName(resultSet.getString("NomeChallenge"));
+                challenges.add(challengeDB);
+            }
+            preparedStatement.close();
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
 
-    ArrayList<PlayerStats> getTopSecondPlace();
+    public void selectAllStats() {
+        ResultSet resultSet;
+        try {
+            PreparedStatement preparedStatement = connection.prepareStatement("SELECT * FROM " + prefix + "Statistic");
+            resultSet = preparedStatement.executeQuery();
+            while (resultSet.next()) {
+                final PlayerStats playerStats = new PlayerStats();
+                playerStats.setPlayerName(resultSet.getString("PlayerName"));
+                playerStats.setNumberOfVictories(resultSet.getInt("NumberVictories"));
+                playerStats.setNumberOfFirstPlace(resultSet.getInt("NumberFirstPlace"));
+                playerStats.setNumberOfSecondPlace(resultSet.getInt("NumberSecondPlace"));
+                playerStats.setNumberOfThirdPlace(resultSet.getInt("NumberThirdPlace"));
+                stats.add(playerStats);
+            }
+            preparedStatement.close();
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
 
-    ArrayList<PlayerStats> getTopThirdPlace();
+    public void selectAllChallengers() {
+        ResultSet resultSet;
+        try {
+            PreparedStatement preparedStatement = connection.prepareStatement("SELECT * FROM " + prefix + "Challenger");
+            resultSet = preparedStatement.executeQuery();
+            while (resultSet.next()) {
+                final Challenger point = new Challenger();
+                point.setPoints(resultSet.getInt("Points"));
+                point.setNomePlayer(resultSet.getString("PlayerName"));
+                playerPoints.add(point);
+            }
+            preparedStatement.close();
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
 
-    void clearAll();
+    public void selectAllDailyWinners() {
+        ResultSet resultSet;
+        try {
+            PreparedStatement preparedStatement = connection.prepareStatement("SELECT * FROM " + prefix + "DailyWinner");
+            resultSet = preparedStatement.executeQuery();
+            while (resultSet.next()) {
+                final DailyWinner dailyWinner = new DailyWinner();
+                dailyWinner.setId(resultSet.getInt("ID"));
+                dailyWinner.setNomeChallenge(resultSet.getString("NomeChallenge"));
+                dailyWinner.setPlayerName(resultSet.getString("PlayerName"));
+                dailyWinner.setReward(resultSet.getString("Reward"));
+                dailyWinners.add(dailyWinner);
+            }
+            preparedStatement.close();
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
 
-    void controlIfChallengeExist(List<String> controlIfChallengeExist);
+    public void selectAllChallengersTopYesterday() {
+        ResultSet resultSet;
+        try {
+            PreparedStatement preparedStatement = connection.prepareStatement("SELECT * FROM " + prefix + "TopYesterday");
+            resultSet = preparedStatement.executeQuery();
+            while (resultSet.next()) {
+                final Challenger point = new Challenger();
+                point.setPoints(resultSet.getInt("Points"));
+                point.setNomePlayer(resultSet.getString("PlayerName"));
+                topYesterday.add(point);
+            }
+            preparedStatement.close();
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
 
-    ArrayList<Challenge> getAllChallenges();
+    public void selectAllOldChallengers() {
+        ResultSet resultSet;
+        try {
+            PreparedStatement preparedStatement = connection.prepareStatement("SELECT * FROM " + prefix + "ChallengerEvent");
+            resultSet = preparedStatement.executeQuery();
+            while (resultSet.next()) {
+                final Challenger point = new Challenger();
+                point.setPoints(resultSet.getInt("Points"));
+                point.setNomePlayer(resultSet.getString("PlayerName"));
+                oldPoints.add(point);
+            }
+            preparedStatement.close();
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
 
-    void clearChallengesFromFile();
+    public void backupDb(int numberOfFiles) {
+        SimpleDateFormat sdf = new SimpleDateFormat("yyyy.MM.dd.HH.mm.ss");
+        Timestamp timestamp = new Timestamp(System.currentTimeMillis());
+        String data = sdf.format(timestamp);
+        File configFile = new File(Main.instance.getDataFolder() +
+                File.separator + "backup", data + ".yml");
+        if (!configFile.exists()) {
+            try {
+                File folder = new File(Main.instance.getDataFolder() +
+                        File.separator + "backup");
+                boolean folderCreate = folder.mkdir();
+                if (!folderCreate) {
+                    if (folder.listFiles().length >= numberOfFiles) {
+                        java.util.Date finalDate = null;
+                        int number = 0;
+                        for (int i = 0; i < folder.listFiles().length; i++) {
+                            try {
+                                Date date = sdf.parse(folder.listFiles()[i].getName());
+                                if (finalDate == null) {
+                                    finalDate = date;
+                                    number = i;
+                                } else {
+                                    int result = finalDate.compareTo(date);
+                                    if (result > 0) {
+                                        finalDate = date;
+                                        number = i;
+                                    }
+                                }
+                            } catch (ParseException e) {
+                                Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+                            }
+                        }
+                        folder.listFiles()[number].delete();
+                    }
+                }
+                configFile.createNewFile();
+                YamlConfiguration file = YamlConfiguration.loadConfiguration(configFile);
+                for (Challenger playerPoint : playerPoints) {
+                    file.set("Points." + playerPoint.getNomePlayer(), playerPoint.getPoints());
+                }
+                for (Challenge challenge : challenges) {
+                    file.set("Challenges." + challenge.getChallengeName(), challenge.getTimeChallenge());
+                }
+                for (DailyWinner dailyWinner : dailyWinners) {
+                    file.set("DailyWinners." + dailyWinner.getId() + ".PlayerName", dailyWinner.getPlayerName());
+                    file.set("DailyWinners." + dailyWinner.getId() + ".NomeChallenge", dailyWinner.getNomeChallenge());
+                    file.set("DailyWinners." + dailyWinner.getId() + ".Reward", dailyWinner.getReward());
+                }
+                for (Challenger challenger : topYesterday) {
+                    file.set("TopYesterday." + challenger.getNomePlayer(), challenger.getPoints());
+                }
+                for (PlayerStats playerStats : stats) {
+                    file.set("Statistic." + playerStats.getPlayerName() + ".NumberVictories", playerStats.getNumberOfVictories());
+                    file.set("Statistic." + playerStats.getPlayerName() + ".NumberFirstPlace", playerStats.getNumberOfFirstPlace());
+                    file.set("Statistic." + playerStats.getPlayerName() + ".NumberSecondPlace", playerStats.getNumberOfSecondPlace());
+                    file.set("Statistic." + playerStats.getPlayerName() + ".NumberThirdPlace", playerStats.getNumberOfThirdPlace());
+                }
+                file.save(configFile);
+            } catch (IOException e) {
+                Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+            }
+        }
+    }
 
-    void clearStats();
+    public void disconnect() {
+        if (connection != null) {
+            try {
+                connection.close();
+                connection = null;
+            } catch (SQLException e) {
+                Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+            }
+        }
+    }
 
-    boolean isPlayerHaveStats(String playerName);
+    public void clearAll() {
+        clearChallenges();
+        clearChallengers();
+        clearChallengersOld();
+        clearDailyWinners();
+        removeTopYesterday();
+        clearStats();
+    }
 
-    void insertChallenge(String challengeName, int time);
+    // Challenges Querys
 
-    void saveOldPointsForChallengeEvents();
+    public String insertDailyChallenges() {
+        int count = 1;
+        if (challenges.isEmpty()) {
+            String nome = "nessuno";
+            List<String> keys = new ArrayList<>(Main.instance.getConfigGesture().getChallenges().keySet());
+            if (Main.instance.getConfigGesture().getChallengeGeneration().equalsIgnoreCase("Random")) {
+                Collections.shuffle(keys);
+            } else if (Main.instance.getConfigGesture().getChallengeGeneration().equalsIgnoreCase("Single")) {
+                Collections.shuffle(keys);
+                Challenge challenge = Main.instance.getConfigGesture().getChallenges().get(keys.get(0));
+                return challenge.getChallengeName();
+            }
+            for (String key : keys) {
+                Challenge challenge = Main.instance.getConfigGesture().getChallenges().get(key);
+                if (count == 1) {
+                    Main.instance.setDailyChallenge(challenge);
+                    nome = challenge.getTypeChallenge();
+                }
+                challenges.add(challenge);
+                count++;
+            }
+            saveChallenges();
+            return nome;
+        } else {
+            for (int i = 0; i < challenges.size(); i++) {
+                if (challenges.get(i).getTimeChallenge() <= 0) {
+                    deleteChallengeWithName(challenges.get(i).getChallengeName());
+                    challenges.remove(i);
+                } else {
+                    if (challenges.get(i).getChallengeName().contains("Event_")) {
+                        Challenge challenge = Main.instance.getConfigGesture().getChallengesEvent().get(challenges.get(i).getChallengeName().replace("Event_", ""));
+                        challenge.setTimeChallenge(challenges.get(i).getTimeChallenge());
+                        Main.instance.setDailyChallenge(challenge);
+                        Bukkit.getServer().getConsoleSender().sendMessage(ChatColor.GREEN + "[Vanilla Challenges] " + challenges.size() + " challenges remain on DB");
+                        return Main.instance.getDailyChallenge().getTypeChallenge();
+                    }
+                    Challenge challenge = Main.instance.getConfigGesture().getChallenges().get(challenges.get(i).getChallengeName());
+                    challenge.setTimeChallenge(challenges.get(i).getTimeChallenge());
+                    Main.instance.setDailyChallenge(challenge);
+                    Bukkit.getServer().getConsoleSender().sendMessage(ChatColor.GREEN + "[Vanilla Challenges] " + challenges.size() + " challenges remain on DB");
+                    return Main.instance.getDailyChallenge().getTypeChallenge();
+                }
+            }
+            return "nessuno";
+        }
+    }
 
-    void resumeOldPoints();
+    // This method has been invoked when the database doesn't have any challenge scheduled
+    public void saveChallenges() {
+        for (Challenge challenge : challenges) {
+            insertChallenge(challenge.getChallengeName(), challenge.getTimeChallenge());
+        }
+    }
 
-    ArrayList<Challenger> getAllOldChallengers();
+    // This method has been invoked when from a command is invoked the adding of a new Event Challenge
+    public void insertChallengeEvent(String challengeName, int time) {
+        try {
+            PreparedStatement preparedStatement = connection.prepareStatement(
+                    "INSERT INTO " + prefix + "Challenge (NomeChallenge,TimeResume) VALUES ('"
+                            + "Event_" + challengeName + "','" + time
+                            + "')");
+            int affectedRows = preparedStatement.executeUpdate();
+            if (affectedRows == 0) {
+                throw new SQLException("Insert failed, no rows affected.");
+            }
+            preparedStatement.close();
+            clearChallengesFromFile();
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
 
-    boolean isChallengePresent(String challengeName);
+    public void clearChallengesFromFile() {
+        List<Challenge> challenges = this.challenges;
+        challenges.add(0, challenges.get(challenges.size() - 1));
+        challenges.remove(challenges.size() - 1);
+        clearChallenges();
+        for (Challenge challenge : challenges) {
+            insertChallenge(challenge.getChallengeName(), challenge.getTimeChallenge());
+        }
+    }
+
+    public void clearChallenges() {
+        PreparedStatement preparedStatement;
+        try {
+            preparedStatement = connection.prepareStatement(
+                    "DROP TABLE " + prefix + "Challenge");
+            preparedStatement.executeUpdate();
+            preparedStatement.close();
+            preparedStatement = connection.prepareStatement(
+                    "CREATE TABLE IF NOT EXISTS " + prefix + "Challenge (`NomeChallenge` VARCHAR(100) NOT NULL PRIMARY KEY, `TimeResume` INT(15) NOT NULL);");
+            preparedStatement.executeUpdate();
+            preparedStatement.close();
+            challenges.clear();
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
+
+    public void clearChallengersOld() {
+        PreparedStatement preparedStatement;
+        try {
+            preparedStatement = connection.prepareStatement(
+                    "DROP TABLE " + prefix + "ChallengerEvent");
+            preparedStatement.executeUpdate();
+            preparedStatement.close();
+            preparedStatement = connection.prepareStatement(
+                    "CREATE TABLE IF NOT EXISTS " + prefix + "ChallengerEvent (`PlayerName` VARCHAR(100) NOT NULL PRIMARY KEY, `Points` INT(15) NOT NULL);");
+            preparedStatement.executeUpdate();
+            preparedStatement.close();
+            oldPoints.clear();
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
+
+    public void deleteChallengeWithName(String challengeName) {
+        try {
+            PreparedStatement preparedStatement = connection
+                    .prepareStatement("DELETE FROM " + prefix + "Challenge WHERE `NomeChallenge`='" + challengeName + "'");
+            preparedStatement.executeUpdate();
+            preparedStatement.close();
+            removeChallenge(challengeName);
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
+
+    public void insertChallenger(String playerName, long points) {
+        try {
+            PreparedStatement preparedStatement = connection.prepareStatement(
+                    "INSERT INTO " + prefix + "Challenger (PlayerName,Points) VALUES ('"
+                            + playerName + "','" + points
+                            + "')");
+            int affectedRows = preparedStatement.executeUpdate();
+            if (affectedRows == 0) {
+                throw new SQLException("Insert failed, no rows affected.");
+            }
+            preparedStatement.close();
+            addPlayerPoints(new Challenger(playerName, points));
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
+
+    public void updateChallenge(String challengeName, int timeResume) {
+        try {
+            PreparedStatement preparedStatement =
+                    connection.prepareStatement("UPDATE " + prefix + "Challenge SET TimeResume = '" + timeResume + "' WHERE NomeChallenge = '" + challengeName + "'");
+            preparedStatement.executeUpdate();
+            preparedStatement.close();
+            updateChallengeTime(challengeName, timeResume);
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
+
+    public void controlIfChallengeExist(List<String> controlIfChallengeExist) {
+        for (String challengeName : controlIfChallengeExist) {
+            for (Challenge challenge : challenges) {
+                if (challenge.getChallengeName().equalsIgnoreCase(challengeName)) {
+                    challenges.remove(challenge);
+                    deleteChallengeWithName(challengeName);
+                    break;
+                }
+            }
+        }
+    }
+
+    // TODO da vedere cosa aggiungere all'array se necessario
+    public void insertChallenge(String challengeName, int timeResume) {
+        try {
+            PreparedStatement preparedStatement = connection.prepareStatement(
+                    "INSERT INTO " + prefix + "Challenge (NomeChallenge,TimeResume) VALUES ('"
+                            + challengeName + "','" + timeResume
+                            + "')");
+            int affectedRows = preparedStatement.executeUpdate();
+            if (affectedRows == 0) {
+                throw new SQLException("Insert failed, no rows affected.");
+            }
+            preparedStatement.close();
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
+
+    public boolean isChallengePresent(String challengeName) {
+        return challenges.stream().anyMatch((Challenge challenge) -> challenge.getChallengeName().equalsIgnoreCase(challengeName));
+    }
+
+    // Players Querys
+
+    public void loadPlayersPoints() {
+        Main.instance.getDailyChallenge().setPlayers(playerPoints);
+        Main.instance.getDailyChallenge().savePoints();
+        List<Challenger> top = Main.instance.getDailyChallenge().getTopPlayers(Main.instance.getConfigGesture().getNumberOfRewardPlayer());
+        int i = 1;
+        while (!top.isEmpty()) {
+            Bukkit.getConsoleSender().sendMessage(
+                    ColorUtils.applyColor(
+                            Main.instance.getConfigGesture().getMessages().get("topPlayers" + i)
+                                    .replace("{number}", "" + i)
+                                    .replace("{player}", top.get(0).getNomePlayer())
+                                    .replace("{points}", MoneyUtils.transform(top.get(0).getPoints()))));
+            top.removeFirst();
+            i++;
+        }
+    }
+
+    public boolean isPresent(String playerName) {
+        return playerPoints.stream().anyMatch((Challenger challenger) -> challenger.getNomePlayer().equalsIgnoreCase(playerName));
+    }
+
+    public void clearChallengers() {
+        try {
+            PreparedStatement preparedStatement = connection.prepareStatement(
+                    "DROP TABLE " + prefix + "Challenger");
+            preparedStatement.executeUpdate();
+            preparedStatement.close();
+            preparedStatement = connection.prepareStatement(
+                    "CREATE TABLE IF NOT EXISTS " + prefix + "Challenger (`PlayerName` VARCHAR(100) NOT NULL PRIMARY KEY, `Points` INT(15) NOT NULL);");
+            preparedStatement.executeUpdate();
+            preparedStatement.close();
+            playerPoints.clear();
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
+
+    public void saveOldPointsForChallengeEvents() {
+        HashMap<String, Long> copyMap = new HashMap<>(Main.instance.getDailyChallenge().getPlayers());
+        for (Map.Entry<String, Long> player : copyMap.entrySet()) {
+            try {
+                if (player.getValue() > 0) {
+                    insertChallengerEvent(player.getKey(), player.getValue());
+                }
+            } catch (Exception ex) {
+                Bukkit.getServer().getConsoleSender().sendMessage("Save Points Event: " + ex.getMessage());
+            }
+        }
+        clearChallengers();
+    }
+
+    public void insertChallengerEvent(String playerName, long points) {
+        try {
+            PreparedStatement preparedStatement = connection.prepareStatement(
+                    "INSERT INTO " + prefix + "ChallengerEvent (PlayerName,Points) VALUES ('"
+                            + playerName + "','" + points
+                            + "')");
+            int affectedRows = preparedStatement.executeUpdate();
+            if (affectedRows == 0) {
+                throw new SQLException("Insert failed, no rows affected.");
+            }
+            preparedStatement.close();
+            oldPoints.add(new Challenger(playerName, points));
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
+
+    public void deleteChallengerWithName(String nomePlayer) {
+        try {
+            PreparedStatement preparedStatement = connection
+                    .prepareStatement("DELETE FROM " + prefix + "Challenger WHERE `PlayerName`='" + nomePlayer + "'");
+            preparedStatement.executeUpdate();
+            preparedStatement.close();
+            removePlayer(nomePlayer);
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
+
+    public void updateChallenger(String playerName, long points) {
+        try {
+            PreparedStatement preparedStatement =
+                    connection.prepareStatement("UPDATE " + prefix + "Challenger SET Points = '" + points + "' WHERE PlayerName = '" + playerName + "'");
+            preparedStatement.executeUpdate();
+            preparedStatement.close();
+            updatePlayer(playerName, points);
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
+
+    public void resumeOldPoints() {
+        clearChallengers();
+        for (Challenger challenger : oldPoints) {
+            insertChallenger(challenger.getNomePlayer(), challenger.getPoints());
+        }
+        clearChallengersOld();
+    }
+
+    // Statistics Querys
+
+    public void insertPlayerStat(PlayerStats playerStats) {
+        try {
+            PreparedStatement preparedStatement = connection.prepareStatement(
+                    "INSERT INTO " + prefix + "Statistic (PlayerName,NumberVictories,NumberFirstPlace,NumberSecondPlace,NumberThirdPlace) VALUES ('"
+                            + playerStats.getPlayerName() + "','" + playerStats.getNumberOfVictories() + "','"
+                            + playerStats.getNumberOfFirstPlace() + "','" + playerStats.getNumberOfSecondPlace() + "','"
+                            + playerStats.getNumberOfThirdPlace()
+                            + "')");
+            int affectedRows = preparedStatement.executeUpdate();
+            if (affectedRows == 0) {
+                throw new SQLException("Insert failed, no rows affected.");
+            }
+            preparedStatement.close();
+            addStat(playerStats);
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
+
+    public void deletePlayerStatWithPlayerName(String playerName) {
+        try {
+            PreparedStatement preparedStatement = connection
+                    .prepareStatement("DELETE FROM " + prefix + "Statistic WHERE `PlayerName`='" + playerName + "'");
+            preparedStatement.executeUpdate();
+            preparedStatement.close();
+            removeStat(playerName);
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
+
+    public void updatePlayerStat(PlayerStats playerStats) {
+        List<PlayerStats> stats = new ArrayList<>(this.stats);
+        while (!stats.isEmpty()) {
+            if (stats.get(0).getPlayerName().equalsIgnoreCase(playerStats.getPlayerName())) {
+                deletePlayerStatWithPlayerName(stats.get(0).getPlayerName());
+                insertPlayerStat(playerStats);
+                return;
+            }
+            stats.removeFirst();
+        }
+        insertPlayerStat(playerStats);
+    }
+
+    public PlayerStats getStatsPlayer(String playerName) {
+        Optional<PlayerStats> player = stats.stream().findFirst().filter((PlayerStats playerStats) -> playerStats.getPlayerName().equalsIgnoreCase(playerName));
+        return player.orElse(null);
+    }
+
+    public List<PlayerStats> getTopVictories() {
+        List<PlayerStats> top = new ArrayList<>(stats);
+        top.sort(Comparator.comparing(PlayerStats::getNumberOfVictories));
+        Collections.reverse(top);
+        return top;
+    }
+
+    public List<PlayerStats> getTopFirstPlace() {
+        List<PlayerStats> top = new ArrayList<>(stats);
+        top.sort(Comparator.comparing(PlayerStats::getNumberOfFirstPlace));
+        Collections.reverse(top);
+        return top;
+    }
+
+    public List<PlayerStats> getTopSecondPlace() {
+        List<PlayerStats> top = new ArrayList<>(stats);
+        top.sort(Comparator.comparing(PlayerStats::getNumberOfSecondPlace));
+        Collections.reverse(top);
+        return top;
+    }
+
+    public List<PlayerStats> getTopThirdPlace() {
+        List<PlayerStats> top = new ArrayList<>(stats);
+        top.sort(Comparator.comparing(PlayerStats::getNumberOfThirdPlace));
+        Collections.reverse(top);
+        return top;
+    }
+
+    public void clearStats() {
+        PreparedStatement preparedStatement;
+        try {
+            preparedStatement = connection.prepareStatement(
+                    "DROP TABLE " + prefix + "Statistic");
+            preparedStatement.executeUpdate();
+            preparedStatement.close();
+            preparedStatement = connection.prepareStatement(
+                    "CREATE TABLE IF NOT EXISTS " + prefix + "Statistic (`PlayerName` VARCHAR(100) NOT NULL PRIMARY KEY, `NumberVictories` INT(10) NOT NULL, `NumberFirstPlace` INT(10) NOT NULL, `NumberSecondPlace` INT(10) NOT NULL, `NumberThirdPlace` INT(10) NOT NULL);");
+            preparedStatement.executeUpdate();
+            preparedStatement.close();
+            stats.clear();
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
+
+    public boolean isPlayerHaveStats(String playerName) {
+        return stats.stream().anyMatch((PlayerStats playerStats) -> playerStats.getPlayerName().equalsIgnoreCase(playerName));
+    }
+
+    // Top Yesterday Querys
+
+    public void removeTopYesterday() {
+        try {
+            PreparedStatement preparedStatement = connection.prepareStatement(
+                    "DROP TABLE " + prefix + "TopYesterday");
+            preparedStatement.executeUpdate();
+            preparedStatement.close();
+            preparedStatement = connection.prepareStatement(
+                    "CREATE TABLE IF NOT EXISTS " + prefix + "TopYesterday (`ID` INT(100) NOT NULL AUTO_INCREMENT PRIMARY KEY, `PlayerName` VARCHAR(100) NOT NULL, `Points` INT(15) NOT NULL);");
+            preparedStatement.executeUpdate();
+            preparedStatement.close();
+            topYesterday.clear();
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
+
+    public void saveTopYesterday(List<Challenger> newTopYesterday) {
+        for (Challenger challenger : newTopYesterday) {
+            insertChallengerTopYesterday(challenger.getNomePlayer(), challenger.getPoints());
+        }
+    }
+
+    public void insertChallengerTopYesterday(String playerName, long points) {
+        try {
+            PreparedStatement preparedStatement = connection.prepareStatement(
+                    "INSERT INTO " + prefix + "TopYesterday (PlayerName,Points) VALUES ('"
+                            + playerName + "','" + points
+                            + "')");
+            int affectedRows = preparedStatement.executeUpdate();
+            if (affectedRows == 0) {
+                throw new SQLException("Insert failed, no rows affected.");
+            }
+            preparedStatement.close();
+            addTopYesterday(new Challenger(playerName, points));
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
+
+    // Daily Top Querys
+
+    public int lastDailyWinnerId() {
+        if (dailyWinners.isEmpty()) {
+            return 1;
+        }
+        return Collections.max(dailyWinners, Comparator.comparing(DailyWinner::getId)).getId();
+    }
+
+    public void insertDailyWinner(DailyWinner dailyWinner) {
+        try {
+            PreparedStatement preparedStatement = connection.prepareStatement(
+                    "INSERT INTO " + prefix + "DailyWinner (NomeChallenge,PlayerName,Reward) VALUES ('"
+                            + dailyWinner.getNomeChallenge() + "','" + dailyWinner.getPlayerName() + "','" + dailyWinner.getReward()
+                            + "')");
+            int affectedRows = preparedStatement.executeUpdate();
+            if (affectedRows == 0) {
+                throw new SQLException("Insert failed, no rows affected.");
+            }
+            preparedStatement.close();
+            addDailyWinner(dailyWinner);
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
+
+    public void deleteDailyWinnerWithId(int id) {
+        try {
+            PreparedStatement preparedStatement = connection
+                    .prepareStatement("DELETE FROM " + prefix + "DailyWinner WHERE `ID`='" + id + "'");
+            preparedStatement.executeUpdate();
+            preparedStatement.close();
+            removeDailyWinner(id);
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
+
+    public void clearDailyWinners() {
+        PreparedStatement preparedStatement;
+        try {
+            preparedStatement = connection.prepareStatement(
+                    "DROP TABLE " + prefix + "DailyWinner");
+            preparedStatement.executeUpdate();
+            preparedStatement.close();
+            preparedStatement = connection.prepareStatement(
+                    "CREATE TABLE IF NOT EXISTS " + prefix + "DailyWinner (`ID` INT(100) NOT NULL AUTO_INCREMENT PRIMARY KEY, `NomeChallenge` VARCHAR(100) NOT NULL, `PlayerName` VARCHAR(100) NOT NULL, `Reward` VARCHAR(100) NOT NULL);");
+            preparedStatement.executeUpdate();
+            preparedStatement.close();
+        } catch (SQLException e) {
+            Logger.log(LogTag.JFR, LogLevel.ERROR, e.getMessage());
+        }
+    }
+
 }
